feat(patients): add tabs in patient view
